FIX not reading in coco format correctly

+++ main.py
@@ -6,8 +6,8 @@ import glob
 from templates import xml_template, object_template
 
 # Edit these
-COCO_DATA_IMAGES = "./sample_data/images"
-COCO_DATA_ANNOTATIONS = "./sample_data/annotations.json"
+COCO_DATA_IMAGES = "./coco/<image-folder>"
+COCO_DATA_ANNOTATIONS = "./coco/<annotations-json>.json"
 OUTPUT_ZIP_NAME = "sample_voc_output"
 
 def prepare_output_dirs():
@@ -30,20 +30,25 @@ def create_pascal_voc_package(data):
         for d in image_data}
         
     for i, (image_id, image_data) in enumerate(imagemap.items()):
-        # print(image_id)
         object_str = ""
         delete_indices = []
+
         for j, k in enumerate(annotation_data):
             if k["image_id"] == image_id:
                 one_image_anno = annotation_data[j]
                 delete_indices.append(j)
-                
+
+                x_min = one_image_anno["bbox"][0]
+                y_min = one_image_anno["bbox"][1]
+                width = one_image_anno["bbox"][2]
+                height = one_image_anno["bbox"][3]
+
                 object_str += object_template.format(
                     labelmap[one_image_anno["category_id"]],
-                    one_image_anno["bbox"][0], # xmin
-                    one_image_anno["bbox"][1], # ymin 
-                    one_image_anno["bbox"][2], # xmax 
-                    one_image_anno["bbox"][3]  # ymax
+                    x_min,
+                    y_min, 
+                    x_min + width, 
+                    y_min + height
                 )
 
 
@@ -53,12 +58,17 @@ def create_pascal_voc_package(data):
         # Refines the list after extracting relevant annotations
         annotation_data = [a for i, a in enumerate(annotation_data) if i not in delete_indices]
 
+        image = os.path.join(COCO_DATA_IMAGES, image_data["fname"])
+        if not os.path.exists(image):
+            print("No image for the particular annotation")
+            continue 
+
         with open(os.path.join(
             "output", OUTPUT_ZIP_NAME, "Annotations", os.path.splitext(image_data["fname"])[0] + ".xml"), "w+") \
                 as f:
             f.write(xml_annotation)
         
-        image = os.path.join(COCO_DATA_IMAGES, image_data["fname"])
+        print(f"Successfully converted {os.path.basename(image)}")
         shutil.copy(image, os.path.join("output", OUTPUT_ZIP_NAME, "JPEGImages", image_data["fname"]))
